Timestamp metrics support (#4)
* Change timestamp from int to int64

* Started implenting exports with timestamps (Tests fail)

* Remove launch.json

* Added .vscode to gitignore

* wip

* wip

* started custom metrics

* wip

* use new costom metrics

* wip

* wip

* Use template struct for custom metrics

* Created tests for new metrics + removed some parameters for buckets

* Fix tests

* Started Timestamp tests (not working yet)

* Add timestamp test

* wip

* Changed collect function to also collect values for gaps

* Added timestamps to readme

* Add cli param for timestamps and windowsize

* Update example conf

+++ dscparser/structs.go
@@ -44,8 +44,8 @@ func (d DSCData) Equals(d2 DSCData) bool {
 type Dataset struct {
 	XMLName       xml.Name        `xml:"array"`
 	Name          string          `xml:"name,attr"`
-	StartTime     int             `xml:"start_time,attr"`
-	StopTime      int             `xml:"stop_time,attr"`
+	StartTime     int64           `xml:"start_time,attr"`
+	StopTime      int64           `xml:"stop_time,attr"`
 	Dimensions    int             `xml:"dimensions,attr"`
 	DimensionInfo []DimensionInfo `xml:"dimension"`
 	Data          Data            `xml:"data"`
